Refactor KasperskyScanEngineClient to separate module and update import in FileValidator
